Added device code grant exception to purge command

+++ src/Console/PurgeCommand.php
@@ -46,7 +46,9 @@ class PurgeCommand extends Command
         Passport::token()->newQuery()->where($constraint)->delete();
         Passport::authCode()->newQuery()->where($constraint)->delete();
         Passport::refreshToken()->newQuery()->where($constraint)->delete();
-        Passport::deviceCode()->newQuery()->where($constraint)->delete();
+        if (Passport::$deviceCodeGrantEnabled) {
+            Passport::deviceCode()->newQuery()->where($constraint)->delete();
+        }
 
         $this->components->info(sprintf('Purged %s.', implode(' and ', array_filter([
             $revoked ? 'revoked items' : null,
